Fix -Wpointer-to-int-cast compiler warnings

+++ lib/gis/spawn.c
@@ -744,32 +744,32 @@ static void parse_argvec(struct spawn *sp, const char **va)
 	    break;
 	}
 	else if (arg == SF_REDIRECT_FILE) {
-	    sp->redirects[sp->num_redirects].dst_fd = NEXT_ARG(va, int);
+	    sp->redirects[sp->num_redirects].dst_fd = (int)NEXT_ARG(va, intptr_t);
 
 	    sp->redirects[sp->num_redirects].src_fd = -1;
-	    sp->redirects[sp->num_redirects].mode = NEXT_ARG(va, int);
+	    sp->redirects[sp->num_redirects].mode = (int)NEXT_ARG(va, intptr_t);
 	    sp->redirects[sp->num_redirects].file = NEXT_ARG(va, const char *);
 
 	    sp->num_redirects++;
 	}
 	else if (arg == SF_REDIRECT_DESCRIPTOR) {
-	    sp->redirects[sp->num_redirects].dst_fd = NEXT_ARG(va, int);
-	    sp->redirects[sp->num_redirects].src_fd = NEXT_ARG(va, int);
+	    sp->redirects[sp->num_redirects].dst_fd = (int)NEXT_ARG(va, intptr_t);
+	    sp->redirects[sp->num_redirects].src_fd = (int)NEXT_ARG(va, intptr_t);
 
 	    sp->redirects[sp->num_redirects].file = NULL;
 	    sp->num_redirects++;
 	}
 	else if (arg == SF_CLOSE_DESCRIPTOR) {
-	    sp->redirects[sp->num_redirects].dst_fd = NEXT_ARG(va, int);
+	    sp->redirects[sp->num_redirects].dst_fd = (int)NEXT_ARG(va, intptr_t);
 
 	    sp->redirects[sp->num_redirects].src_fd = -1;
 	    sp->redirects[sp->num_redirects].file = NULL;
 	    sp->num_redirects++;
 	}
 	else if (arg == SF_SIGNAL) {
-	    sp->signals[sp->num_signals].which = NEXT_ARG(va, int);
-	    sp->signals[sp->num_signals].action = NEXT_ARG(va, int);
-	    sp->signals[sp->num_signals].signum = NEXT_ARG(va, int);
+	    sp->signals[sp->num_signals].which = (int)NEXT_ARG(va, intptr_t);
+	    sp->signals[sp->num_signals].action = (int)NEXT_ARG(va, intptr_t);
+	    sp->signals[sp->num_signals].signum = (int)NEXT_ARG(va, intptr_t);
 
 	    sp->signals[sp->num_signals].valid = 0;
 	    sp->num_signals++;
